+++ constants/constants.go
@@ -1,0 +1,6 @@
+package constants
+
+const (
+	numTask    = 50
+	formatDate = "20060102"
+)
